1. Code organization for easier analysis 2. Separate code into separate files for each GUI and class
